+++ src/main/resources/db/migration/V1__Init_db.sql
@@ -25,7 +25,7 @@ DROP TABLE IF EXISTS `customers`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `customers` (
-  `id` bigint(20) NOT NULL,
+  `id` int(20) NOT NULL,
   `customer_name` varchar(255) DEFAULT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -40,7 +40,7 @@ DROP TABLE IF EXISTS `hibernate_sequences`;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `hibernate_sequences` (
   `sequence_name` varchar(255) NOT NULL,
-  `sequence_next_hi_value` bigint(20) DEFAULT NULL,
+  `sequence_next_hi_value` int(20) DEFAULT NULL,
   PRIMARY KEY (`sequence_name`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 /*!40101 SET character_set_client = @saved_cs_client */;
@@ -53,7 +53,7 @@ DROP TABLE IF EXISTS `note_objects`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `note_objects` (
-  `note_obj_id` bigint(20) NOT NULL AUTO_INCREMENT,
+  `note_obj_id` int(20) NOT NULL AUTO_INCREMENT,
   `note_obj_content` varchar(255) DEFAULT NULL,
   `note_type` int(11) DEFAULT NULL,
   PRIMARY KEY (`note_obj_id`)
@@ -68,8 +68,8 @@ DROP TABLE IF EXISTS `notes`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `notes` (
-  `note_id` bigint(20) NOT NULL AUTO_INCREMENT,
-  `note_creator_id` bigint(20) DEFAULT NULL,
+  `note_id` int(20) NOT NULL AUTO_INCREMENT,
+  `note_creator_id` int(20) DEFAULT NULL,
   PRIMARY KEY (`note_id`),
   KEY `FK2t9ngfw0wpgb82lc1v2po6ouq` (`note_creator_id`),
   CONSTRAINT `FK2t9ngfw0wpgb82lc1v2po6ouq` FOREIGN KEY (`note_creator_id`) REFERENCES `users` (`id`)
@@ -84,8 +84,8 @@ DROP TABLE IF EXISTS `notes_note_assigned_to`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `notes_note_assigned_to` (
-  `note_note_id` bigint(20) NOT NULL,
-  `note_assigned_to_id` bigint(20) NOT NULL,
+  `note_note_id` int(20) NOT NULL,
+  `note_assigned_to_id` int(20) NOT NULL,
   UNIQUE KEY `UK_n35pxvgpobidkyfmh268epi10` (`note_assigned_to_id`),
   KEY `FK2jn234rn3jo1gsvuafckt814d` (`note_note_id`),
   CONSTRAINT `FK2jn234rn3jo1gsvuafckt814d` FOREIGN KEY (`note_note_id`) REFERENCES `notes` (`note_id`)
@@ -100,8 +100,8 @@ DROP TABLE IF EXISTS `notes_note_data`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `notes_note_data` (
-  `note_note_id` bigint(20) NOT NULL,
-  `note_data_note_obj_id` bigint(20) NOT NULL,
+  `note_note_id` int(20) NOT NULL,
+  `note_data_note_obj_id` int(20) NOT NULL,
   UNIQUE KEY `UK_feivrmdmkyl0enh0kt0wc25f0` (`note_data_note_obj_id`),
   KEY `FKoistn2fpm3np9luojjhggy7pn` (`note_note_id`),
   CONSTRAINT `FKmjn45gja35kmi0ss03bb9g79t` FOREIGN KEY (`note_data_note_obj_id`) REFERENCES `note_objects` (`note_obj_id`),
@@ -117,10 +117,10 @@ DROP TABLE IF EXISTS `projects`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `projects` (
-  `id` bigint(20) NOT NULL,
+  `id` int(20) NOT NULL,
   `project_description` varchar(255) DEFAULT NULL,
   `project_title` varchar(255) DEFAULT NULL,
-  `project_customer_id` bigint(20) DEFAULT NULL,
+  `project_customer_id` int(20) DEFAULT NULL,
   PRIMARY KEY (`id`),
   KEY `FKfraqgwq0afgegsyuv7k9rmeg1` (`project_customer_id`),
   CONSTRAINT `FKfraqgwq0afgegsyuv7k9rmeg1` FOREIGN KEY (`project_customer_id`) REFERENCES `customers` (`id`)
@@ -135,8 +135,8 @@ DROP TABLE IF EXISTS `projects_project_sprints`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `projects_project_sprints` (
-  `project_id` bigint(20) NOT NULL,
-  `project_sprints_id` bigint(20) NOT NULL,
+  `project_id` int(20) NOT NULL,
+  `project_sprints_id` int(20) NOT NULL,
   UNIQUE KEY `UK_mc1sr77g4qv6jou22r6cswv6m` (`project_sprints_id`),
   KEY `FKchbtcgbsb9m0wkx9fvvy72nsj` (`project_id`),
   CONSTRAINT `FKchbtcgbsb9m0wkx9fvvy72nsj` FOREIGN KEY (`project_id`) REFERENCES `projects` (`id`),
@@ -152,8 +152,8 @@ DROP TABLE IF EXISTS `projects_project_tasks`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `projects_project_tasks` (
-  `project_id` bigint(20) NOT NULL,
-  `project_tasks_id` bigint(20) NOT NULL,
+  `project_id` int(20) NOT NULL,
+  `project_tasks_id` int(20) NOT NULL,
   UNIQUE KEY `UK_q26ij2ymdos33rcqc6bhtlhyu` (`project_tasks_id`),
   KEY `FKbhqsi6st5fx6tcw45ttcxtsp9` (`project_id`),
   CONSTRAINT `FKbhqsi6st5fx6tcw45ttcxtsp9` FOREIGN KEY (`project_id`) REFERENCES `projects` (`id`),
@@ -169,8 +169,8 @@ DROP TABLE IF EXISTS `projects_project_users`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `projects_project_users` (
-  `project_id` bigint(20) NOT NULL,
-  `project_users_id` bigint(20) NOT NULL,
+  `project_id` int(20) NOT NULL,
+  `project_users_id` int(20) NOT NULL,
   UNIQUE KEY `UK_ed0lj06vn12cgb7wvjkrhkna3` (`project_users_id`),
   KEY `FK5gaise1h4hp6c6dtokimvryco` (`project_id`),
   CONSTRAINT `FK5gaise1h4hp6c6dtokimvryco` FOREIGN KEY (`project_id`) REFERENCES `projects` (`id`),
@@ -186,7 +186,7 @@ DROP TABLE IF EXISTS `sprints`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `sprints` (
-  `id` bigint(20) NOT NULL,
+  `id` int(20) NOT NULL,
   `sprint_end_date` datetime DEFAULT NULL,
   `sprint_start_date` datetime DEFAULT NULL,
   `sprint_title` varchar(255) DEFAULT NULL,
@@ -202,8 +202,8 @@ DROP TABLE IF EXISTS `sprints_sprint_tasks`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `sprints_sprint_tasks` (
-  `sprint_id` bigint(20) NOT NULL,
-  `sprint_tasks_id` bigint(20) NOT NULL,
+  `sprint_id` int(20) NOT NULL,
+  `sprint_tasks_id` int(20) NOT NULL,
   UNIQUE KEY `UK_d5sl37kunpilxb5g35bqll1vf` (`sprint_tasks_id`),
   KEY `FKf4ifpde4dljl0f4xafphgxsg4` (`sprint_id`),
   CONSTRAINT `FK8rg5ynnqx06733oxi6pkaaw4v` FOREIGN KEY (`sprint_tasks_id`) REFERENCES `tasks` (`id`),
@@ -219,7 +219,7 @@ DROP TABLE IF EXISTS `tasks`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `tasks` (
-  `id` bigint(20) NOT NULL,
+  `id` int(20) NOT NULL,
   `task_description` varchar(255) DEFAULT NULL,
   `task_end_date` datetime DEFAULT NULL,
   `task_estimated_timeh` int(11) NOT NULL,
@@ -238,8 +238,8 @@ DROP TABLE IF EXISTS `tasks_task_assigned_users`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `tasks_task_assigned_users` (
-  `task_id` bigint(20) NOT NULL,
-  `task_assigned_users_id` bigint(20) NOT NULL,
+  `task_id` int(20) NOT NULL,
+  `task_assigned_users_id` int(20) NOT NULL,
   UNIQUE KEY `UK_91ahac9lr4q60273xw7excg3p` (`task_assigned_users_id`),
   KEY `FKp0joj940wfsaftfdmovl0lgrj` (`task_id`),
   CONSTRAINT `FKbqyu05wrmenl4duc0upu3vvg2` FOREIGN KEY (`task_assigned_users_id`) REFERENCES `users` (`id`),
@@ -255,7 +255,7 @@ DROP TABLE IF EXISTS `users`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!40101 SET character_set_client = utf8 */;
 CREATE TABLE `users` (
-  `id` bigint(20) NOT NULL,
+  `id` int(20) NOT NULL,
   `email` varchar(255) DEFAULT NULL,
   `password` varchar(255) DEFAULT NULL,
   `phone` varchar(255) DEFAULT NULL,
